\name{TGLG-package}
\alias{TGLG-package}
\alias{TGLG}
\docType{package}
\title{
\packageTitle{TGLG}
}
\description{
\packageDescription{TGLG}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TGLG}
\packageIndices{TGLG}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TGLG}

Maintainer: \packageMaintainer{TGLG}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
